[Refactor]: 컴포넌트 EditProfile 디렉토리 밑으로 이동
