+++ src/components/common/EditProfile/ProfileSubmitButton.jsx
@@ -1,12 +1,12 @@
 import React, { useEffect } from "react";
 import { useSelector } from "react-redux";
 import axios from "axios";
-import { BASE_URL } from "../../utils/URL";
-import { Button } from "./Button";
+import { BASE_URL } from "../../../utils/URL";
+import { Button } from "../Button";
 import { useMediaQuery } from "react-responsive";
 import styled from "styled-components";
-import { PlaceRightRow } from "../../styles/utils";
-import { SmallText } from "../../styles/text";
+import { PlaceRightRow } from "../../../styles/utils";
+import { SmallText } from "../../../styles/text";
 
 const StyledProfileSubmitButton = styled(PlaceRightRow)`
   @media (min-height: 768px) {
